@@ -11,41 +11,41 @@
   },
   "dependencies": {
     "@react-native-community/blur": "3.6.0",
+    "@react-native-community/cameraroll": "4.0.1",
     "@react-native-community/masked-view": "0.1.10",
     "@react-native-community/netinfo": "5.9.9",
     "@react-navigation/native": "5.8.10",
     "@react-navigation/stack": "5.12.8",
     "react": "16.13.1",
-    "react-native": "0.63.3",
+    "react-native": "0.63.4",
     "react-native-document-picker": "4.1.0",
     "react-native-fs": "2.16.6",
     "react-native-gesture-handler": "1.9.0",
-    "react-native-get-random-values": "1.5.0",
     "react-native-haptic-feedback": "1.11.0",
-    "react-native-image-crop-picker": "0.35.1",
-    "react-native-reanimated": "2.0.0-alpha.9.2",
+    "react-native-image-crop-picker": "0.35.2",
+    "react-native-reanimated": "2.0.0-rc.0",
     "react-native-safe-area-context": "3.1.9",
-    "react-native-screens": "2.15.0",
+    "react-native-screens": "2.16.1",
     "react-native-share": "4.1.0",
     "react-native-svg": "12.1.0",
-    "stream-chat": "2.8.1",
+    "stream-chat": "2.9.0",
     "stream-chat-react-native": "link:../../native-package"
   },
   "devDependencies": {
-    "@babel/core": "7.12.7",
+    "@babel/core": "7.12.9",
     "@babel/runtime": "7.12.5",
     "@react-native-community/eslint-config": "2.0.0",
-    "@types/jest": "26.0.15",
-    "@types/react-native": "0.63.36",
+    "@types/jest": "26.0.16",
+    "@types/react-native": "0.63.40",
     "@types/react-test-renderer": "16.9.3",
-    "@typescript-eslint/eslint-plugin": "4.8.1",
-    "@typescript-eslint/parser": "4.8.1",
+    "@typescript-eslint/eslint-plugin": "4.9.0",
+    "@typescript-eslint/parser": "4.9.0",
     "babel-jest": "26.6.3",
-    "eslint": "7.14.0",
+    "eslint": "7.15.0",
     "jest": "26.6.3",
     "metro-react-native-babel-preset": "0.64.0",
     "react-test-renderer": "16.13.1",
-    "typescript": "4.1.2"
+    "typescript": "4.1.3"
   },
   "jest": {
     "preset": "react-native",
