@@ -130,25 +130,25 @@
     "@react-native-community/netinfo": "5.9.9",
     "@testing-library/jest-native": "3.4.3",
     "@testing-library/react-native": "7.1.0",
-    "@types/eslint": "7.2.5",
-    "@types/jest": "26.0.15",
+    "@types/eslint": "7.2.6",
+    "@types/jest": "26.0.19",
     "@types/lodash": "4.14.165",
     "@types/mime-types": "2.1.0",
     "@types/react": "16.9.56",
-    "@types/react-native": "0.63.36",
+    "@types/react-native": "0.63.40",
     "@types/react-native-actionsheet": "2.4.2",
     "@types/react-test-renderer": "16.9.3",
-    "@types/styled-components": "5.1.4",
-    "@typescript-eslint/eslint-plugin": "4.8.1",
-    "@typescript-eslint/parser": "4.8.1",
+    "@types/styled-components": "5.1.5",
+    "@typescript-eslint/eslint-plugin": "4.8.2",
+    "@typescript-eslint/parser": "4.8.2",
     "babel-eslint": "10.1.0",
     "babel-jest": "26.6.3",
-    "babel-loader": "8.2.1",
+    "babel-loader": "8.2.2",
     "babel-plugin-i18next-extract": "0.8.2",
     "babel-plugin-module-resolver": "4.0.0",
-    "babel-plugin-react-native-web": "0.14.8",
-    "babel-plugin-styled-components": "1.11.1",
-    "eslint": "7.14.0",
+    "babel-plugin-react-native-web": "0.14.9",
+    "babel-plugin-styled-components": "1.12.0",
+    "eslint": "7.15.0",
     "eslint-config-prettier": "6.15.0",
     "eslint-plugin-babel": "5.3.1",
     "eslint-plugin-better-styled-components": "1.1.2",
