@@ -1,33 +1,197 @@
 import * as React from 'react';
 import { useTheme } from '../../contexts/themeContext/ThemeContext';
 
+import { Audio } from '../../icons/Audio';
 import { CSV } from '../../icons/CSV';
 import { DOC } from '../../icons/DOC';
+import { DOCX } from '../../icons/DOCX';
+import { GenericFile } from '../../icons/GenericFile';
+import { HTML } from '../../icons/HTML';
+import { MD } from '../../icons/MD';
+import { ODT } from '../../icons/ODT';
 import { PDF } from '../../icons/PDF';
 import { PPT } from '../../icons/PPT';
+import { PPTX } from '../../icons/PPTX';
+import { RAR } from '../../icons/RAR';
+import { RTF } from '../../icons/RTF';
+import { SEVEN_Z } from '../../icons/SEVEN_Z';
 import { TAR } from '../../icons/TAR';
+import { TXT } from '../../icons/TXT';
 import { XLS } from '../../icons/XLS';
+import { XLSX } from '../../icons/XLSX';
 import { ZIP } from '../../icons/ZIP';
 
 import type { IconProps } from '../../icons/utils/base';
 
+// https://www.iana.org/assignments/media-types/media-types.xhtml#audio
+const audioFileTypes = [
+  'audio/1d-interleaved-parityfec',
+  'audio/32kadpcm',
+  'audio/3gpp',
+  'audio/3gpp2',
+  'audio/aac',
+  'audio/ac3',
+  'audio/AMR',
+  'audio/AMR-WB',
+  'audio/amr-wb+',
+  'audio/aptx',
+  'audio/asc',
+  'audio/ATRAC-ADVANCED-LOSSLESS',
+  'audio/ATRAC-X',
+  'audio/ATRAC3',
+  'audio/basic',
+  'audio/BV16',
+  'audio/BV32',
+  'audio/clearmode',
+  'audio/CN',
+  'audio/DAT12',
+  'audio/dls',
+  'audio/dsr-es201108',
+  'audio/dsr-es202050',
+  'audio/dsr-es202211',
+  'audio/dsr-es202212',
+  'audio/DV',
+  'audio/DVI4',
+  'audio/eac3',
+  'audio/encaprtp',
+  'audio/EVRC',
+  'audio/EVRC-QCP',
+  'audio/EVRC0',
+  'audio/EVRC1',
+  'audio/EVRCB',
+  'audio/EVRCB0',
+  'audio/EVRCB1',
+  'audio/EVRCNW',
+  'audio/EVRCNW0',
+  'audio/EVRCNW1',
+  'audio/EVRCWB',
+  'audio/EVRCWB0',
+  'audio/EVRCWB1',
+  'audio/EVS',
+  'audio/example',
+  'audio/flexfec',
+  'audio/fwdred',
+  'audio/G711-0',
+  'audio/G719',
+  'audio/G7221',
+  'audio/G722',
+  'audio/G723',
+  'audio/G726-16',
+  'audio/G726-24',
+  'audio/G726-32',
+  'audio/G726-40',
+  'audio/G728',
+  'audio/G729',
+  'audio/G7291',
+  'audio/G729D',
+  'audio/G729E',
+  'audio/GSM',
+  'audio/GSM-EFR',
+  'audio/GSM-HR-08',
+  'audio/iLBC',
+  'audio/ip-mr_v2.5',
+  'audio/L8',
+  'audio/L16',
+  'audio/L20',
+  'audio/L24',
+  'audio/LPC',
+  'audio/MELP',
+  'audio/MELP600',
+  'audio/MELP1200',
+  'audio/MELP2400',
+  'audio/mhas',
+  'audio/mobile-xmf',
+  'audio/MPA',
+  'audio/mp4',
+  'audio/MP4A-LATM',
+  'audio/mpa-robust',
+  'audio/mpeg',
+  'audio/mpeg4-generic',
+  'audio/ogg',
+  'audio/opus',
+  'audio/parityfec',
+  'audio/PCMA',
+  'audio/PCMA-WB',
+  'audio/PCMU',
+  'audio/PCMU-WB',
+  'audio/prs.sid',
+  'audio/raptorfec',
+  'audio/RED',
+  'audio/rtp-enc-aescm128',
+  'audio/rtploopback',
+  'audio/rtp-midi',
+  'audio/rtx',
+  'audio/SMV',
+  'audio/SMV0',
+  'audio/SMV-QCP',
+  'audio/sofa',
+  'audio/sp-midi',
+  'audio/speex',
+  'audio/t140c',
+  'audio/t38',
+  'audio/telephone-event',
+  'audio/TETRA_ACELP',
+  'audio/TETRA_ACELP_BB',
+  'audio/tone',
+  'audio/TSVCIS',
+  'audio/UEMCLIP',
+  'audio/ulpfec',
+  'audio/usac',
+  'audio/VDVI',
+  'audio/VMR-WB',
+  'audio/vnd.3gpp.iufp',
+  'audio/vnd.4SB',
+  'audio/vnd.audiokoz',
+  'audio/vnd.CELP',
+  'audio/vnd.cisco.nse',
+  'audio/vnd.cmles.radio-events',
+  'audio/vnd.cns.anp1',
+  'audio/vnd.cns.inf1',
+  'audio/vnd.dece.audio',
+  'audio/vnd.digital-winds',
+  'audio/vnd.dlna.adts',
+  'audio/vnd.dolby.heaac.1',
+  'audio/vnd.dolby.heaac.2',
+  'audio/vnd.dolby.mlp',
+  'audio/vnd.dolby.mps',
+  'audio/vnd.dolby.pl2',
+  'audio/vnd.dolby.pl2x',
+  'audio/vnd.dolby.pl2z',
+  'audio/vnd.dolby.pulse.1',
+  'audio/vnd.dra',
+  'audio/vnd.dts',
+  'audio/vnd.dts.hd',
+  'audio/vnd.dts.uhd',
+  'audio/vnd.dvb.file',
+  'audio/vnd.everad.plj',
+  'audio/vnd.hns.audio',
+  'audio/vnd.lucent.voice',
+  'audio/vnd.ms-playready.media.pya',
+  'audio/vnd.nokia.mobile-xmf',
+  'audio/vnd.nortel.vbk',
+  'audio/vnd.nuera.ecelp4800',
+  'audio/vnd.nuera.ecelp7470',
+  'audio/vnd.nuera.ecelp9600',
+  'audio/vnd.octel.sbc',
+  'audio/vnd.presonus.multitrack',
+  'audio/vnd.qcelp',
+  'audio/vnd.rhetorex.32kadpcm',
+  'audio/vnd.rip',
+  'audio/vnd.sealedmedia.softseal.mpeg',
+  'audio/vnd.vmx.cvsd',
+  'audio/vorbis',
+  'audio/vorbis-config',
+];
+
 // Partially based on:
 // https://stackoverflow.com/a/4212908/2570866
 
-const wordMimeTypes = [
+const docMimeTypes = [
   // Microsoft Word
   // .doc .dot
   'application/msword',
   // .doc .dot
   'application/msword-template',
-  // .docx
-  'application/vnd.openxmlformats-officedocument.wordprocessingml.document',
-  // .dotx (no test)
-  'application/vnd.openxmlformats-officedocument.wordprocessingml.template',
-  // .docm
-  'application/vnd.ms-word.document.macroEnabled.12',
-  // .dotm (no test)
-  'application/vnd.ms-word.template.macroEnabled.12',
 
   // LibreOffice/OpenOffice Writer
   // .odt
@@ -40,12 +204,33 @@ const wordMimeTypes = [
   // NOTE: firefox doesn't know mimetype so maybe ignore
 ];
 
-const excelMimeTypes = [
-  // TODO: maybe more data files
+const docXMimeTypes = [
+  // Microsoft Word
+  // .docx
+  'application/vnd.openxmlformats-officedocument.wordprocessingml.document',
+  // .dotx (no test)
+  'application/vnd.openxmlformats-officedocument.wordprocessingml.template',
+  // .docm
+  'application/vnd.ms-word.document.macroEnabled.12',
+  // .dotm (no test)
+  'application/vnd.ms-word.template.macroEnabled.12',
+];
 
+const excelMimeTypes = [
   // Microsoft Excel
   // .xls .xlt .xla (no test for .xla)
   'application/vnd.ms-excel',
+
+  // LibreOffice/OpenOffice Calc
+  // .ods
+  'application/vnd.oasis.opendocument.spreadsheet',
+  // .ots
+  'application/vnd.oasis.opendocument.spreadsheet-template',
+  // .fods
+  'application/vnd.oasis.opendocument.spreadsheet-flat-xml',
+];
+
+const excelXMimeTypes = [
   // .xlsx
   'application/vnd.openxmlformats-officedocument.spreadsheetml.sheet',
   // .xltx (no test)
@@ -58,22 +243,35 @@ const excelMimeTypes = [
   'application/vnd.ms-excel.addin.macroEnabled.12',
   // .xlsb (no test)
   'application/vnd.ms-excel.addin.macroEnabled.12',
+];
 
-  // LibreOffice/OpenOffice Calc
-  // .ods
-  'application/vnd.oasis.opendocument.spreadsheet',
-  // .ots
-  'application/vnd.oasis.opendocument.spreadsheet-template',
-  // .fods
-  'application/vnd.oasis.opendocument.spreadsheet-flat-xml',
-  // .uos
-  // NOTE: firefox doesn't know mimetype so maybe ignore
+const odtMimeTypes = [
+  // LibreOffice/OpenOffice Writer
+  // .odt
+  'application/vnd.oasis.opendocument.text',
+  // .ott
+  'application/vnd.oasis.opendocument.text-template',
+  // .fodt
+  'application/vnd.oasis.opendocument.text-flat-xml',
 ];
 
 const powerpointMimeTypes = [
   // Microsoft Word
   // .ppt .pot .pps .ppa (no test for .ppa)
   'application/vnd.ms-powerpoint',
+
+  // LibreOffice/OpenOffice Writer
+  // .odp
+  'application/vnd.oasis.opendocument.presentation',
+  // .otp
+  'application/vnd.oasis.opendocument.presentation-template',
+  // .fodp
+  'application/vnd.oasis.opendocument.presentation-flat-xml',
+  // .uop
+  // NOTE: firefox doesn't know mimetype so maybe ignore
+];
+
+const powerpointXMimeTypes = [
   // .pptx
   'application/vnd.openxmlformats-officedocument.presentationml.presentation',
   // .potx (no test)
@@ -88,21 +286,9 @@ const powerpointMimeTypes = [
   'application/vnd.ms-powerpoint.template.macroEnabled.12',
   // .ppsm
   'application/vnd.ms-powerpoint.slideshow.macroEnabled.12',
-
-  // LibreOffice/OpenOffice Writer
-  // .odp
-  'application/vnd.oasis.opendocument.presentation',
-  // .otp
-  'application/vnd.oasis.opendocument.presentation-template',
-  // .fodp
-  'application/vnd.oasis.opendocument.presentation-flat-xml',
-  // .uop
-  // NOTE: firefox doesn't know mimetype so maybe ignore
 ];
 
 const tarFileTypes = [
-  // .z7
-  'application/x-7z-compressed',
   // .ar
   'application/x-archive',
   // .tar
@@ -125,43 +311,6 @@ const tarFileTypes = [
   'application/x-xz',
   // .war
   'application/x-webarchive',
-  // .rar
-  'application/vnd.rar',
-];
-
-const codeFileTypes = [
-  // .html .htm
-  'text/html',
-  // .css
-  'text/css',
-  // .js
-  'application/x-javascript',
-  // .json
-  'application/json',
-  // .py
-  'text/x-python',
-  // .go
-  'text/x-go',
-  // .c
-  'text/x-csrc',
-  // .cpp
-  'text/x-c++src',
-  // .rb
-  'application/x-ruby',
-  // .rust
-  'text/rust',
-  // .java
-  'text/x-java',
-  // .php
-  'application/x-php',
-  // .cs
-  'text/x-csharp',
-  // .scala
-  'text/x-scala',
-  // .erl
-  'text/x-erlang',
-  // .sh
-  'application/x-shellscript',
 ];
 
 const zipFileTypes = [
@@ -172,28 +321,50 @@ const zipFileTypes = [
 ];
 
 const mimeTypeToIconMap: Record<string, React.FC<IconProps>> = {
-  'application/pdf': PDF,
+  'application/pdf': PDF, // .pdf
+  'application/rtf': RTF, // .rtf
+  'application/vnd.rar': RAR, // .rar
+  'application/x-7z-compressed': SEVEN_Z, // .z7
   'text/csv': CSV, // .csv
+  'text/html': HTML, // .html .htm
+  'text/markdown': MD, // .md
+  'text/plain': TXT, // .txt
 };
 
-for (const type of wordMimeTypes) {
+for (const type of audioFileTypes) {
+  mimeTypeToIconMap[type] = Audio;
+}
+
+for (const type of docMimeTypes) {
   mimeTypeToIconMap[type] = DOC;
+}
+
+for (const type of docXMimeTypes) {
+  mimeTypeToIconMap[type] = DOCX;
 }
 
 for (const type of excelMimeTypes) {
   mimeTypeToIconMap[type] = XLS;
 }
 
+for (const type of excelXMimeTypes) {
+  mimeTypeToIconMap[type] = XLSX;
+}
+
+for (const type of odtMimeTypes) {
+  mimeTypeToIconMap[type] = ODT;
+}
+
 for (const type of powerpointMimeTypes) {
   mimeTypeToIconMap[type] = PPT;
 }
 
-for (const type of tarFileTypes) {
-  mimeTypeToIconMap[type] = TAR;
+for (const type of powerpointXMimeTypes) {
+  mimeTypeToIconMap[type] = PPTX;
 }
 
-for (const type of codeFileTypes) {
-  mimeTypeToIconMap[type] = DOC;
+for (const type of tarFileTypes) {
+  mimeTypeToIconMap[type] = TAR;
 }
 
 for (const type of zipFileTypes) {
@@ -201,12 +372,12 @@ for (const type of zipFileTypes) {
 }
 
 function mimeTypeToIcon(mimeType?: string): React.FC<IconProps> {
-  if (!mimeType) return DOC;
+  if (!mimeType) return GenericFile;
 
   const Icon = mimeTypeToIconMap[mimeType];
   if (Icon) return Icon;
 
-  return DOC;
+  return GenericFile;
 }
 
 export type FileIconProps = {
