@@ -4,7 +4,12 @@ import React, {
   useEffect,
   useState,
 } from 'react';
-import { KeyboardAvoidingViewProps, StyleSheet, Text } from 'react-native';
+import {
+  KeyboardAvoidingViewProps,
+  StyleSheet,
+  Text,
+  View,
+} from 'react-native';
 import debounce from 'lodash/debounce';
 import throttle from 'lodash/throttle';
 import Immutable from 'seamless-immutable';
@@ -42,6 +47,14 @@ import { MessageReplies as MessageRepliesDefault } from '../Message/MessageSimpl
 import { MessageSimple as MessageSimpleDefault } from '../Message/MessageSimple/MessageSimple';
 import { MessageStatus as MessageStatusDefault } from '../Message/MessageSimple/MessageStatus';
 import { ReactionList as ReactionListDefault } from '../Message/MessageSimple/ReactionList';
+import { AttachButton as AttachButtonDefault } from '../MessageInput/AttachButton';
+import { CommandsButton as CommandsButtonDefault } from '../MessageInput/CommandsButton';
+import { FileUploadPreview as FileUploadPreviewDefault } from '../MessageInput/FileUploadPreview';
+import { ImageUploadPreview as ImageUploadPreviewDefault } from '../MessageInput/ImageUploadPreview';
+import { MoreOptionsButton as MoreOptionsButtonDefault } from '../MessageInput/MoreOptionsButton';
+import { SendButton as SendButtonDefault } from '../MessageInput/SendButton';
+import { UploadProgressIndicator as UploadProgressIndicatorDefault } from '../MessageInput/UploadProgressIndicator';
+import { Reply as ReplyDefault } from '../Reply/Reply';
 
 import {
   ChannelContextValue,
@@ -52,11 +65,18 @@ import {
   useChatContext,
 } from '../../contexts/chatContext/ChatContext';
 import {
+  InputMessageInputContextValue,
+  MessageInputProvider,
+} from '../../contexts/messageInputContext/MessageInputContext';
+import {
   ActionProps,
   MessagesContextValue,
   MessagesProvider,
 } from '../../contexts/messagesContext/MessagesContext';
-import { SuggestionsProvider } from '../../contexts/suggestionsContext/SuggestionsContext';
+import {
+  SuggestionsContextValue,
+  SuggestionsProvider,
+} from '../../contexts/suggestionsContext/SuggestionsContext';
 import { useTheme } from '../../contexts/themeContext/ThemeContext';
 import {
   ThreadContextValue,
@@ -100,6 +120,8 @@ export type ChannelPropsWithContext<
   >
 > &
   Pick<ChatContextValue<At, Ch, Co, Ev, Me, Re, Us>, 'client'> &
+  Partial<InputMessageInputContextValue<At, Ch, Co, Ev, Me, Re, Us>> &
+  Partial<SuggestionsContextValue<Co, Us>> &
   Pick<TranslationContextValue, 't'> &
   Partial<
     Pick<
@@ -131,6 +153,7 @@ export type ChannelPropsWithContext<
       | 'MessageStatus'
       | 'MessageText'
       | 'ReactionList'
+      | 'Reply'
       | 'supportedReactions'
       | 'UrlPreview'
     >
@@ -221,7 +244,9 @@ export const ChannelWithContext = <
 ) => {
   const {
     additionalKeyboardAvoidingViewProps,
+    additionalTextInputProps,
     additionalTouchableProps,
+    AttachButton = AttachButtonDefault,
     Attachment = AttachmentDefault,
     AttachmentActions = AttachmentActionsDefault,
     AttachmentFileIcon = FileIconDefault,
@@ -232,25 +257,37 @@ export const ChannelWithContext = <
     channel,
     children,
     client,
+    closeSuggestions,
+    CommandsButton = CommandsButtonDefault,
+    compressImageQuality,
     disableIfFrozenChannel = true,
     disableKeyboardCompatibleView = false,
     disableTypingIndicator,
     dismissKeyboardOnMessageTouch = true,
+    doDocUploadRequest,
+    doImageUploadRequest,
     doMarkReadRequest,
     doSendMessageRequest,
     doUpdateMessageRequest,
     EmptyStateIndicator = EmptyStateIndicatorDefault,
     FileAttachment = FileAttachmentDefault,
     FileAttachmentGroup = FileAttachmentGroupDefault,
+    FileUploadPreview = FileUploadPreviewDefault,
     formatDate,
     Gallery = GalleryDefault,
     Giphy = GiphyDefault,
+    hasFilePicker = true,
+    hasImagePicker = true,
+    ImageUploadPreview = ImageUploadPreviewDefault,
+    initialValue,
+    Input,
     keyboardBehavior,
     KeyboardCompatibleView = KeyboardCompatibleViewDefault,
     keyboardVerticalOffset,
     LoadingErrorIndicator = LoadingErrorIndicatorDefault,
     LoadingIndicator = LoadingIndicatorDefault,
     markdownRules,
+    maxNumberOfFiles = 10,
     Message = MessageDefault,
     MessageAvatar = MessageAvatarDefault,
     MessageContent = MessageContentDefault,
@@ -261,11 +298,21 @@ export const ChannelWithContext = <
     MessageSimple = MessageSimpleDefault,
     MessageStatus = MessageStatusDefault,
     MessageText,
+    MoreOptionsButton = MoreOptionsButtonDefault,
+    numberOfLines = 5,
+    onChangeText,
+    openSuggestions,
     ReactionList = ReactionListDefault,
+    Reply = ReplyDefault,
+    SendButton = SendButtonDefault,
+    sendImageAsync = false,
+    setInputRef,
     StickyHeader,
     supportedReactions = reactionDataDefault,
     t,
     thread: threadProps,
+    updateSuggestions,
+    UploadProgressIndicator = UploadProgressIndicatorDefault,
     UrlPreview = CardDefault,
   } = props;
 
@@ -314,6 +361,9 @@ export const ChannelWithContext = <
   const [read, setRead] = useState<
     ChannelContextValue<At, Ch, Co, Ev, Me, Re, Us>['read']
   >({} as ChannelContextValue<At, Ch, Co, Ev, Me, Re, Us>['read']);
+  const [replyTo, setReplyTo] = useState<
+    boolean | MessageType<At, Ch, Co, Ev, Me, Re, Us>
+  >(false);
   const [thread, setThread] = useState<
     ThreadContextValue<At, Ch, Co, Ev, Me, Re, Us>['thread']
   >(threadProps || null);
@@ -634,7 +684,7 @@ export const ChannelWithContext = <
     }
   };
 
-  const sendMessage: MessagesContextValue<
+  const sendMessage: InputMessageInputContextValue<
     At,
     Ch,
     Co,
@@ -734,7 +784,7 @@ export const ChannelWithContext = <
     trailing: true,
   });
 
-  const editMessage: MessagesContextValue<
+  const editMessage: InputMessageInputContextValue<
     At,
     Ch,
     Co,
@@ -759,7 +809,19 @@ export const ChannelWithContext = <
     setEditing(message);
   };
 
-  const clearEditingState: MessagesContextValue<
+  const setReplyToState: MessagesContextValue<
+    At,
+    Ch,
+    Co,
+    Ev,
+    Me,
+    Re,
+    Us
+  >['setReplyToState'] = (message) => {
+    setReplyTo(message);
+  };
+
+  const clearEditingState: InputMessageInputContextValue<
     At,
     Ch,
     Co,
@@ -768,6 +830,16 @@ export const ChannelWithContext = <
     Re,
     Us
   >['clearEditingState'] = () => setEditing(false);
+
+  const clearReplyToState: InputMessageInputContextValue<
+    At,
+    Ch,
+    Co,
+    Ev,
+    Me,
+    Re,
+    Us
+  >['clearReplyToState'] = () => setReplyTo(false);
 
   const removeMessage: MessagesContextValue<
     At,
@@ -896,6 +968,43 @@ export const ChannelWithContext = <
     watchers,
   };
 
+  const messageInputContext: InputMessageInputContextValue<
+    At,
+    Ch,
+    Co,
+    Ev,
+    Me,
+    Re,
+    Us
+  > = {
+    additionalTextInputProps,
+    AttachButton,
+    clearEditingState,
+    clearReplyToState,
+    CommandsButton,
+    compressImageQuality,
+    doDocUploadRequest,
+    doImageUploadRequest,
+    editing,
+    editMessage,
+    FileUploadPreview,
+    hasFilePicker,
+    hasImagePicker,
+    ImageUploadPreview,
+    initialValue,
+    Input,
+    maxNumberOfFiles,
+    MoreOptionsButton,
+    numberOfLines,
+    onChangeText,
+    replyTo,
+    SendButton,
+    sendImageAsync,
+    sendMessage,
+    setInputRef,
+    UploadProgressIndicator,
+  };
+
   const messagesContext: MessagesContextValue<At, Ch, Co, Ev, Me, Re, Us> = {
     ...actionProps,
     additionalTouchableProps,
@@ -906,11 +1015,8 @@ export const ChannelWithContext = <
     CardCover,
     CardFooter,
     CardHeader,
-    clearEditingState,
     disableTypingIndicator,
     dismissKeyboardOnMessageTouch,
-    editing,
-    editMessage,
     FileAttachment,
     FileAttachmentGroup,
     formatDate,
@@ -933,12 +1039,19 @@ export const ChannelWithContext = <
     MessageText,
     ReactionList,
     removeMessage,
+    Reply,
     retrySendMessage,
-    sendMessage,
     setEditingState,
+    setReplyToState,
     supportedReactions,
     updateMessage,
     UrlPreview,
+  };
+
+  const suggestionsContext: Partial<SuggestionsContextValue<Co, Us>> = {
+    closeSuggestions,
+    openSuggestions,
+    updateSuggestions,
   };
 
   const threadContext: ThreadContextValue<At, Ch, Co, Ev, Me, Re, Us> = {
@@ -981,7 +1094,13 @@ export const ChannelWithContext = <
       <ChannelProvider<At, Ch, Co, Ev, Me, Re, Us> value={channelContext}>
         <MessagesProvider<At, Ch, Co, Ev, Me, Re, Us> value={messagesContext}>
           <ThreadProvider<At, Ch, Co, Ev, Me, Re, Us> value={threadContext}>
-            <SuggestionsProvider<Co, Us>>{children}</SuggestionsProvider>
+            <SuggestionsProvider<Co, Us> value={suggestionsContext}>
+              <MessageInputProvider<At, Ch, Co, Ev, Me, Re, Us>
+                value={messageInputContext}
+              >
+                <View style={{ height: '100%' }}>{children}</View>
+              </MessageInputProvider>
+            </SuggestionsProvider>
           </ThreadProvider>
         </MessagesProvider>
       </ChannelProvider>
