@@ -127,6 +127,7 @@ export type MessagePropsWithContext<
     | 'reactionsEnabled'
     | 'retrySendMessage'
     | 'setEditingState'
+    | 'setReplyToState'
     | 'supportedReactions'
     | 'updateMessage'
   > &
@@ -246,6 +247,7 @@ const MessageWithContext = <
     setData,
     setEditingState,
     setOverlay,
+    setReplyToState,
     showAvatar,
     showMessageStatus,
     supportedReactions,
@@ -333,11 +335,16 @@ const MessageWithContext = <
             } else if (
               cur.type === 'image' &&
               !cur.title_link &&
-              !cur.og_scrape_url &&
-              (cur.image_url || cur.thumb_url)
+              !cur.og_scrape_url
             ) {
-              acc.images.push(cur);
-              acc.other = []; // remove other attachments if an image exists
+              /**
+               * this next if is not combined with the above one for cases where we have
+               * an image with no url links at all falling back to being an attachment
+               */
+              if (cur.image_url || cur.thumb_url) {
+                acc.images.push(cur);
+                acc.other = []; // remove other attachments if an image exists
+              }
               // only add other attachments if there are no files/images
             } else if (!acc.files.length && !acc.images.length) {
               acc.other.push(cur);
@@ -519,7 +526,7 @@ const MessageWithContext = <
     const reply = {
       action: () => {
         setOverlay('none');
-        onOpenThread();
+        setReplyToState(message);
       },
       icon: <CurveLineLeftUp />,
       title: t('Reply'),
@@ -657,7 +664,7 @@ const MessageWithContext = <
     [onLongPressMessage],
   );
 
-  return (
+  return message.deleted_at || messageContentOrder.length ? (
     <TapGestureHandler
       enabled={animatedLongPress}
       maxDurationMs={3000}
@@ -669,7 +676,7 @@ const MessageWithContext = <
         </MessageProvider>
       </Animated.View>
     </TapGestureHandler>
-  );
+  ) : null;
 };
 
 const areEqual = <
@@ -719,11 +726,19 @@ const areEqual = <
     prevMessage.user?.banned === nextMessage.user?.banned;
   if (!messageUserBannedEqual) return false;
 
+  const prevMessageAttachments = prevMessage.attachments;
+  const nextMessageAttachments = nextMessage.attachments;
   const attachmentsEqual =
-    (Array.isArray(prevMessage.attachments) &&
-      Array.isArray(nextMessage.attachments) &&
-      prevMessage.attachments.length === nextMessage.attachments.length) ||
-    prevMessage.attachments === nextMessage.attachments;
+    (Array.isArray(prevMessageAttachments) &&
+      Array.isArray(nextMessageAttachments) &&
+      prevMessageAttachments.length === nextMessageAttachments.length &&
+      prevMessageAttachments.every((attachment, index) =>
+        attachment.type === 'image'
+          ? attachment.image_url === nextMessageAttachments[index].image_url &&
+            attachment.thumb_url === nextMessageAttachments[index].thumb_url
+          : attachment.type === nextMessageAttachments[index].type,
+      )) ||
+    prevMessageAttachments === nextMessageAttachments;
   if (!attachmentsEqual) return false;
 
   const latestReactionsEqual =
@@ -805,6 +820,7 @@ export const Message = <
     removeMessage,
     retrySendMessage,
     setEditingState,
+    setReplyToState,
     supportedReactions,
     updateMessage,
   } = useMessagesContext<At, Ch, Co, Ev, Me, Re, Us>();
@@ -832,6 +848,7 @@ export const Message = <
         setData,
         setEditingState,
         setOverlay,
+        setReplyToState,
         supportedReactions,
         t,
         updateMessage,
