@@ -1,36 +1,56 @@
 import React from 'react';
+import { TouchableOpacity } from 'react-native-gesture-handler';
+
+import { useAttachmentPickerContext } from '../../contexts/attachmentPickerContext/AttachmentPickerContext';
 import {
-  GestureResponderEvent,
-  StyleSheet,
-  TouchableOpacity,
-} from 'react-native';
-import { Attach } from '../../icons';
-
+  ChannelContextValue,
+  useChannelContext,
+} from '../../contexts/channelContext/ChannelContext';
 import { useTheme } from '../../contexts/themeContext/ThemeContext';
+import { Attach } from '../../icons/Attach';
 
-const styles = StyleSheet.create({
-  container: {
-    marginBottom: 8,
-    marginRight: 8,
-  },
-});
+import type { GestureResponderEvent } from 'react-native';
 
-export type AttachButtonProps = {
-  /** Disables the button */
-  disabled?: boolean;
-  /** Function that opens an attachment action sheet */
+import type {
+  DefaultAttachmentType,
+  DefaultChannelType,
+  DefaultCommandType,
+  DefaultEventType,
+  DefaultMessageType,
+  DefaultReactionType,
+  DefaultUserType,
+  UnknownType,
+} from '../../types/types';
+
+type AttachButtonPropsWithContext<
+  At extends UnknownType = DefaultAttachmentType,
+  Ch extends UnknownType = DefaultChannelType,
+  Co extends string = DefaultCommandType,
+  Ev extends UnknownType = DefaultEventType,
+  Me extends UnknownType = DefaultMessageType,
+  Re extends UnknownType = DefaultReactionType,
+  Us extends UnknownType = DefaultUserType
+> = Pick<ChannelContextValue<At, Ch, Co, Ev, Me, Re, Us>, 'disabled'> & {
+  /** Function that opens attachment options bottom sheet */
   handleOnPress?: (event: GestureResponderEvent) => void;
+  selectedPicker?: 'images';
 };
 
-/**
- * UI Component for attach button in MessageInput component.
- *
- * @example ./AttachButton.md
- */
-export const AttachButton: React.FC<AttachButtonProps> = (props) => {
-  const { disabled = false, handleOnPress } = props;
+const AttachButtonWithContext = <
+  At extends UnknownType = DefaultAttachmentType,
+  Ch extends UnknownType = DefaultChannelType,
+  Co extends string = DefaultCommandType,
+  Ev extends UnknownType = DefaultEventType,
+  Me extends UnknownType = DefaultMessageType,
+  Re extends UnknownType = DefaultReactionType,
+  Us extends UnknownType = DefaultUserType
+>(
+  props: AttachButtonPropsWithContext<At, Ch, Co, Ev, Me, Re, Us>,
+) => {
+  const { disabled, handleOnPress, selectedPicker } = props;
   const {
     theme: {
+      colors: { primary, textGrey },
       messageInput: { attachButton },
     },
   } = useTheme();
@@ -39,12 +59,84 @@ export const AttachButton: React.FC<AttachButtonProps> = (props) => {
     <TouchableOpacity
       disabled={disabled}
       onPress={handleOnPress}
-      style={[styles.container, attachButton]}
+      style={[attachButton]}
       testID='attach-button'
     >
-      <Attach height={24} width={24} />
+      <Attach pathFill={selectedPicker === 'images' ? primary : textGrey} />
     </TouchableOpacity>
   );
+};
+
+const areEqual = <
+  At extends UnknownType = DefaultAttachmentType,
+  Ch extends UnknownType = DefaultChannelType,
+  Co extends string = DefaultCommandType,
+  Ev extends UnknownType = DefaultEventType,
+  Me extends UnknownType = DefaultMessageType,
+  Re extends UnknownType = DefaultReactionType,
+  Us extends UnknownType = DefaultUserType
+>(
+  prevProps: AttachButtonPropsWithContext<At, Ch, Co, Ev, Me, Re, Us>,
+  nextProps: AttachButtonPropsWithContext<At, Ch, Co, Ev, Me, Re, Us>,
+) => {
+  const {
+    disabled: prevDisabled,
+    handleOnPress: prevHandleOnPress,
+    selectedPicker: prevSelectedPicker,
+  } = prevProps;
+  const {
+    disabled: nextDisabled,
+    handleOnPress: nextHandleOnPress,
+    selectedPicker: nextSelectedPicker,
+  } = nextProps;
+
+  const disabledEqual = prevDisabled === nextDisabled;
+  if (!disabledEqual) return false;
+
+  const handleOnPressEqual = prevHandleOnPress === nextHandleOnPress;
+  if (!handleOnPressEqual) return false;
+
+  const selectedPickerEqual = prevSelectedPicker === nextSelectedPicker;
+  if (!selectedPickerEqual) return false;
+
+  return true;
+};
+
+const MemoizedAttachButton = React.memo(
+  AttachButtonWithContext,
+  areEqual,
+) as typeof AttachButtonWithContext;
+
+export type AttachButtonProps<
+  At extends UnknownType = DefaultAttachmentType,
+  Ch extends UnknownType = DefaultChannelType,
+  Co extends string = DefaultCommandType,
+  Ev extends UnknownType = DefaultEventType,
+  Me extends UnknownType = DefaultMessageType,
+  Re extends UnknownType = DefaultReactionType,
+  Us extends UnknownType = DefaultUserType
+> = Partial<AttachButtonPropsWithContext<At, Ch, Co, Ev, Me, Re, Us>>;
+
+/**
+ * UI Component for attach button in MessageInput component.
+ *
+ * @example ./AttachButton.md
+ */
+export const AttachButton = <
+  At extends UnknownType = DefaultAttachmentType,
+  Ch extends UnknownType = DefaultChannelType,
+  Co extends string = DefaultCommandType,
+  Ev extends UnknownType = DefaultEventType,
+  Me extends UnknownType = DefaultMessageType,
+  Re extends UnknownType = DefaultReactionType,
+  Us extends UnknownType = DefaultUserType
+>(
+  props: AttachButtonProps<At, Ch, Co, Ev, Me, Re, Us>,
+) => {
+  const { disabled = false } = useChannelContext<At, Ch, Co, Ev, Me, Re, Us>();
+  const { selectedPicker } = useAttachmentPickerContext();
+
+  return <MemoizedAttachButton {...{ disabled, selectedPicker }} {...props} />;
 };
 
 AttachButton.displayName = 'AttachButton{messageInput}';
