@@ -1,5 +1,6 @@
 import React, { useEffect } from 'react';
 import { StyleSheet, Text, View } from 'react-native';
+import Svg, { Defs, LinearGradient, Rect, Stop } from 'react-native-svg';
 
 import { Message as DefaultMessage } from '../Message/Message';
 import {
@@ -15,8 +16,7 @@ import { useChannelContext } from '../../contexts/channelContext/ChannelContext'
 import { useTheme } from '../../contexts/themeContext/ThemeContext';
 import { useThreadContext } from '../../contexts/threadContext/ThreadContext';
 import { useTranslationContext } from '../../contexts/translationContext/TranslationContext';
-
-import type { Message as StreamMessage } from 'stream-chat';
+import { vw } from '../../utils/utils';
 
 import type {
   DefaultAttachmentType,
@@ -30,12 +30,21 @@ import type {
 } from '../../types/types';
 
 const styles = StyleSheet.create({
+  messagePadding: {
+    paddingHorizontal: 8,
+  },
   newThread: {
     alignItems: 'center',
-    backgroundColor: '#F4F9FF',
-    borderRadius: 4,
-    margin: 10,
-    padding: 8,
+    height: 24,
+    justifyContent: 'center',
+    marginTop: 8,
+  },
+  text: {
+    fontSize: 12,
+    textAlign: 'center',
+  },
+  threadHeaderContainer: {
+    marginVertical: 8,
   },
 });
 
@@ -125,8 +134,14 @@ export const Thread = <
 
   const {
     theme: {
+      colors: { background, textGrey },
       thread: {
-        newThread: { text, ...newThread },
+        newThread: {
+          backgroundGradientStart,
+          backgroundGradientStop,
+          text,
+          ...newThread
+        },
       },
     },
   } = useTheme();
@@ -165,35 +180,78 @@ export const Thread = <
 
   if (!thread) return null;
 
-  const headerComponent = (
-    <>
-      <DefaultMessage<At, Ch, Co, Ev, Me, Re, Us>
-        alignment={'left'}
-        groupStyles={['single']}
-        message={thread}
-        preventPress
-        threadList
-      />
-      <View style={[styles.newThread, newThread]}>
-        <Text style={text}>{t('Start of a new thread')}</Text>
+  const numberOfReplies = thread.reply_count;
+
+  const footerComponent = (
+    <View style={styles.threadHeaderContainer}>
+      <View style={styles.messagePadding}>
+        <DefaultMessage<At, Ch, Co, Ev, Me, Re, Us>
+          alignment={'left'}
+          groupStyles={['single']}
+          message={thread}
+          preventPress
+          threadList
+        />
       </View>
-    </>
+      <View style={[styles.newThread, newThread]}>
+        <Svg
+          height={newThread.height ?? 24}
+          style={{ position: 'absolute' }}
+          width={vw(100)}
+        >
+          <Rect
+            fill='url(#gradient)'
+            height={newThread.height ?? 24}
+            width={vw(100)}
+            x={0}
+            y={0}
+          />
+          <Defs>
+            <LinearGradient
+              gradientUnits='userSpaceOnUse'
+              id='gradient'
+              x1={0}
+              x2={0}
+              y1={0}
+              y2={newThread.height ?? 24}
+            >
+              <Stop
+                offset={1}
+                stopColor={backgroundGradientStart || '#F7F7F7'}
+                stopOpacity={1}
+              />
+              <Stop
+                offset={0}
+                stopColor={backgroundGradientStop || background}
+                stopOpacity={1}
+              />
+            </LinearGradient>
+          </Defs>
+        </Svg>
+        <Text style={[styles.text, { color: textGrey }, text]}>
+          {numberOfReplies === 1
+            ? t('{{ numberOfReplies }} Reply', {
+                numberOfReplies,
+              })
+            : t('{{ numberOfReplies }} Replies', {
+                numberOfReplies,
+              })}
+        </Text>
+      </View>
+    </View>
   );
 
   return (
     <React.Fragment key={`thread-${thread.id}-${channel?.cid || ''}`}>
       <MessageList<At, Ch, Co, Ev, Me, Re, Us>
-        HeaderComponent={headerComponent}
+        FooterComponent={footerComponent}
         threadList
         {...additionalMessageListProps}
       />
       <MessageInput<At, Ch, Co, Ev, Me, Re, Us>
         additionalTextInputProps={{ autoFocus, editable: !disabled }}
-        parent_id={thread.id as StreamMessage<At, Me, Us>['parent_id']}
         {...additionalMessageInputProps}
       />
     </React.Fragment>
   );
 };
-
-Thread.displayName = 'Thread{thread{newThread}}';
