@@ -7,7 +7,6 @@ import type { DebouncedFunc } from 'lodash';
 import type {
   ChannelState,
   MessageResponse,
-  StreamChat,
   Message as StreamMessage,
 } from 'stream-chat';
 
@@ -30,6 +29,7 @@ import type { MessageTextProps } from '../../components/Message/MessageSimple/Me
 import type { MarkdownRules } from '../../components/Message/MessageSimple/utils/renderText';
 import type { Message } from '../../components/MessageList/hooks/useMessageList';
 import type { ReactionListProps } from '../../components/Message/MessageSimple/ReactionList';
+import type { ReplyProps } from '../../components/Reply/Reply';
 import type { TDateTimeParserInput } from '../../contexts/translationContext/TranslationContext';
 import type {
   DefaultAttachmentType,
@@ -64,18 +64,6 @@ export type MessageWithDates<
   readBy: boolean | number;
 };
 
-export const isEditingBoolean = <
-  At extends UnknownType = DefaultAttachmentType,
-  Ch extends UnknownType = DefaultChannelType,
-  Co extends string = DefaultCommandType,
-  Ev extends UnknownType = DefaultEventType,
-  Me extends UnknownType = DefaultMessageType,
-  Re extends UnknownType = DefaultReactionType,
-  Us extends UnknownType = DefaultUserType
->(
-  editing: MessagesContextValue<At, Ch, Co, Ev, Me, Re, Us>['editing'],
-): editing is boolean => typeof editing === 'boolean';
-
 export type MessagesContextValue<
   At extends UnknownType = DefaultAttachmentType,
   Ch extends UnknownType = DefaultChannelType,
@@ -105,11 +93,8 @@ export type MessagesContextValue<
    * Defaults to https://github.com/GetStream/stream-chat-react-native/blob/master/src/components/Attachment/Card.tsx
    */
   Card: React.ComponentType<CardProps<At>>;
-  clearEditingState: () => void;
   /** Should keyboard be dismissed when messaged is touched */
   dismissKeyboardOnMessageTouch: boolean;
-  editing: boolean | Message<At, Ch, Co, Ev, Me, Re, Us>;
-  editMessage: StreamChat<At, Ch, Co, Ev, Me, Re, Us>['updateMessage'];
   /**
    * Custom UI component to display File type attachment.
    * Defaults to https://github.com/GetStream/stream-chat-react-native/blob/master/src/components/Attachment/FileAttachment.tsx
@@ -179,11 +164,17 @@ export type MessagesContextValue<
     id: string;
     parent_id?: StreamMessage<At, Me, Us>['parent_id'];
   }) => void;
+  /**
+   * Custom UI component for reply component.
+   *
+   * Defaults to and accepts same props as: [Reply](https://getstream.github.io/stream-chat-react-native/#reply)
+   */
+  Reply: React.ComponentType<ReplyProps<At, Ch, Co, Ev, Me, Re, Us>>;
   retrySendMessage: (
     message: MessageResponse<At, Ch, Co, Me, Re, Us>,
   ) => Promise<void>;
-  sendMessage: (message: Partial<StreamMessage<At, Me, Us>>) => Promise<void>;
   setEditingState: (message: Message<At, Ch, Co, Ev, Me, Re, Us>) => void;
+  setReplyToState: (message: Message<At, Ch, Co, Ev, Me, Re, Us>) => void;
   supportedReactions: ReactionData[];
   updateMessage: (
     updatedMessage: MessageResponse<At, Ch, Co, Me, Re, Us>,
@@ -270,7 +261,6 @@ const areEqual = <
       additionalTouchableProps: prevAdditionalTouchableProps,
       disableTypingIndicator: prevDisableTypingIndicator,
       dismissKeyboardOnMessageTouch: prevDismissKeyboardOnMessageTouch,
-      editing: prevEditing,
       hasMore: prevHasMore,
       loadingMore: prevLoadingMore,
       markdownRules: prevMarkdownRules,
@@ -284,7 +274,6 @@ const areEqual = <
       additionalTouchableProps: nextAdditionalTouchableProps,
       disableTypingIndicator: nextDisableTypingIndicator,
       dismissKeyboardOnMessageTouch: nextDismissKeyboardOnMessageTouch,
-      editing: nextEditing,
       hasMore: nextHasMore,
       loadingMore: nextLoadingMore,
       markdownRules: nextMarkdownRules,
@@ -307,9 +296,6 @@ const areEqual = <
 
   const loadingMoreEqual = prevLoadingMore === nextLoadingMore;
   if (!loadingMoreEqual) return false;
-
-  const editingEqual = !!prevEditing === !!nextEditing;
-  if (!editingEqual) return false;
 
   const messageContentOrderEqual =
     prevMessageContentOrder.length === nextMessageContentOrder.length &&
@@ -387,8 +373,8 @@ export const useMessagesContext = <
   >;
 
 /**
- * Typescript currently does not support partial inference so if ChatContext
- * typing is desired while using the HOC withChannelContext the Props for the
+ * Typescript currently does not support partial inference so if MessagesContext
+ * typing is desired while using the HOC withMessagesContext the Props for the
  * wrapped component must be provided as the first generic.
  */
 export const withMessagesContext = <
