@@ -2,6 +2,7 @@ import React, {
   PropsWithChildren,
   useContext,
   useEffect,
+  useRef,
   useState,
 } from 'react';
 import {
@@ -18,6 +19,10 @@ import Animated, {
   withTiming,
 } from 'react-native-reanimated';
 
+import {
+  AttachmentPickerContextValue,
+  AttachmentPickerProvider,
+} from '../attachmentPickerContext/AttachmentPickerContext';
 import { ChannelInfoOverlayProvider } from '../channelInfoOverlayContext/ChannelInfoOverlayContext';
 import { ImageGalleryProvider } from '../imageGalleryContext/ImageGalleryContext';
 import { MessageOverlayProvider } from '../messageOverlayContext/MessageOverlayContext';
@@ -28,6 +33,17 @@ import {
 } from '../translationContext/TranslationContext';
 import { getDisplayName } from '../utils/getDisplayName';
 
+import {
+  AttachmentPicker,
+  AttachmentPickerProps,
+} from '../../components/AttachmentPicker/AttachmentPicker';
+import { AttachmentPickerError as DefaultAttachmentPickerError } from '../../components/AttachmentPicker/components/AttachmentPickerError';
+import { AttachmentPickerBottomSheetHandle as DefaultAttachmentPickerBottomSheetHandle } from '../../components/AttachmentPicker/components/AttachmentPickerBottomSheetHandle';
+import { AttachmentPickerErrorImage as DefaultAttachmentPickerErrorImage } from '../../components/AttachmentPicker/components/AttachmentPickerErrorImage';
+import { CameraSelectorIcon as DefaultCameraSelectorIcon } from '../../components/AttachmentPicker/components/CameraSelectorIcon';
+import { FileSelectorIcon as DefaultFileSelectorIcon } from '../../components/AttachmentPicker/components/FileSelectorIcon';
+import { ImageOverlaySelectedComponent as DefaultImageOverlaySelectedComponent } from '../../components/AttachmentPicker/components/ImageOverlaySelectedComponent';
+import { ImageSelectorIcon as DefaultImageSelectorIcon } from '../../components/AttachmentPicker/components/ImageSelectorIcon';
 import { ChannelInfoOverlay } from '../../components/ChannelInfoOverlay/ChannelInfoOverlay';
 import {
   ImageGallery,
@@ -36,6 +52,8 @@ import {
 import { MessageOverlay } from '../../components/MessageOverlay/MessageOverlay';
 import { BlurView } from '../../native';
 import { useStreami18n } from '../../utils/useStreami18n';
+
+import type BottomSheet from '@gorhom/bottom-sheet';
 
 import type { Theme } from '../themeContext/utils/theme';
 
@@ -50,6 +68,7 @@ import type {
   UnknownType,
 } from '../../types/types';
 import type { Streami18n } from '../../utils/Streami18n';
+import type { BottomSheetMethods } from '@gorhom/bottom-sheet/lib/typescript/types';
 
 export type BlurType = 'light' | 'dark' | undefined;
 
@@ -80,10 +99,24 @@ export const OverlayContext = React.createContext<OverlayContextValue>(
 );
 
 type Props<Us extends UnknownType = DefaultUserType> = PropsWithChildren<
-  ImageGalleryCustomComponents<Us> & {
-    i18nInstance?: Streami18n;
-    value?: Partial<OverlayContextValue>;
-  }
+  Partial<AttachmentPickerProps> &
+    Partial<
+      Pick<
+        AttachmentPickerContextValue,
+        | 'attachmentPickerBottomSheetHeight'
+        | 'attachmentSelectionBarHeight'
+        | 'bottomInset'
+        | 'CameraSelectorIcon'
+        | 'FileSelectorIcon'
+        | 'ImageSelectorIcon'
+      >
+    > &
+    ImageGalleryCustomComponents<Us> & {
+      closePicker?: (ref: React.RefObject<BottomSheetMethods>) => void;
+      i18nInstance?: Streami18n;
+      openPicker?: (ref: React.RefObject<BottomSheetMethods>) => void;
+      value?: Partial<OverlayContextValue>;
+    }
 >;
 
 const WildcardDefault: React.FC<{ visible: boolean }> = () => null;
@@ -99,7 +132,45 @@ export const OverlayProvider = <
 >(
   props: Props<Us>,
 ) => {
-  const { children, i18nInstance, imageGalleryCustomComponents, value } = props;
+  const {
+    AttachmentPickerBottomSheetHandle = DefaultAttachmentPickerBottomSheetHandle,
+    attachmentPickerBottomSheetHandleHeight,
+    attachmentPickerBottomSheetHeight,
+    AttachmentPickerError = DefaultAttachmentPickerError,
+    attachmentPickerErrorButtonText,
+    AttachmentPickerErrorImage = DefaultAttachmentPickerErrorImage,
+    attachmentPickerErrorText,
+    attachmentSelectionBarHeight,
+    bottomInset,
+    CameraSelectorIcon = DefaultCameraSelectorIcon,
+    children,
+    closePicker = (ref) => ref.current?.close(),
+    FileSelectorIcon = DefaultFileSelectorIcon,
+    i18nInstance,
+    imageGalleryCustomComponents,
+    ImageOverlaySelectedComponent = DefaultImageOverlaySelectedComponent,
+    ImageSelectorIcon = DefaultImageSelectorIcon,
+    numberOfAttachmentImagesToLoadPerCall,
+    numberOfAttachmentPickerImageColumns,
+    openPicker = (ref) => ref.current?.snapTo(0),
+    value,
+  } = props;
+
+  const attachmentPickerProps = {
+    AttachmentPickerBottomSheetHandle,
+    attachmentPickerBottomSheetHandleHeight,
+    attachmentPickerBottomSheetHeight,
+    AttachmentPickerError,
+    attachmentPickerErrorButtonText,
+    AttachmentPickerErrorImage,
+    attachmentPickerErrorText,
+    attachmentSelectionBarHeight,
+    ImageOverlaySelectedComponent,
+    numberOfAttachmentImagesToLoadPerCall,
+    numberOfAttachmentPickerImageColumns,
+  };
+
+  const bottomSheetRef = useRef<BottomSheet>(null);
 
   const [translators, setTranslators] = useState<TranslationContextValue>({
     t: (key: string) => key,
@@ -137,6 +208,7 @@ export const OverlayProvider = <
   }, [overlay]);
 
   useEffect(() => {
+    bottomSheetRef.current?.close();
     cancelAnimation(overlayOpacity);
     if (overlay !== 'none') {
       overlayOpacity.value = withTiming(1);
@@ -147,6 +219,17 @@ export const OverlayProvider = <
 
   // Setup translators
   useStreami18n({ i18nInstance, setTranslators });
+
+  const attachmentPickerContext = {
+    attachmentPickerBottomSheetHeight,
+    attachmentSelectionBarHeight,
+    bottomInset,
+    CameraSelectorIcon,
+    closePicker: () => closePicker(bottomSheetRef),
+    FileSelectorIcon,
+    ImageSelectorIcon,
+    openPicker: () => openPicker(bottomSheetRef),
+  };
 
   const overlayStyle = useAnimatedStyle<ViewStyle>(
     () => ({
@@ -171,34 +254,40 @@ export const OverlayProvider = <
       <OverlayContext.Provider value={overlayContext}>
         <ChannelInfoOverlayProvider<At, Ch, Co, Ev, Me, Re, Us>>
           <MessageOverlayProvider<At, Ch, Co, Ev, Me, Re, Us>>
-            <ImageGalleryProvider>
-              {children}
-              <ThemeProvider style={overlayContext.style}>
-                <Animated.View
-                  pointerEvents={overlay === 'none' ? 'none' : 'auto'}
-                  style={[StyleSheet.absoluteFill, overlayStyle]}
-                >
-                  <BlurView
-                    blurType={blurType}
-                    style={[StyleSheet.absoluteFill, { height, width }]}
+            <AttachmentPickerProvider value={attachmentPickerContext}>
+              <ImageGalleryProvider>
+                {children}
+                <ThemeProvider style={overlayContext.style}>
+                  <Animated.View
+                    pointerEvents={overlay === 'none' ? 'none' : 'auto'}
+                    style={[StyleSheet.absoluteFill, overlayStyle]}
+                  >
+                    <BlurView
+                      blurType={blurType}
+                      style={[StyleSheet.absoluteFill, { height, width }]}
+                    />
+                  </Animated.View>
+                  {Wildcard && <Wildcard visible={overlay === 'wildcard'} />}
+                  <ChannelInfoOverlay<At, Ch, Co, Ev, Me, Re, Us>
+                    overlayOpacity={overlayOpacity}
+                    visible={overlay === 'channelInfo'}
                   />
-                </Animated.View>
-                {Wildcard && <Wildcard visible={overlay === 'wildcard'} />}
-                <ChannelInfoOverlay<At, Ch, Co, Ev, Me, Re, Us>
-                  overlayOpacity={overlayOpacity}
-                  visible={overlay === 'channelInfo'}
-                />
-                <MessageOverlay<At, Ch, Co, Ev, Me, Re, Us>
-                  overlayOpacity={overlayOpacity}
-                  visible={overlay === 'message'}
-                />
-                <ImageGallery<At, Ch, Co, Ev, Me, Re, Us>
-                  imageGalleryCustomComponents={imageGalleryCustomComponents}
-                  overlayOpacity={overlayOpacity}
-                  visible={overlay === 'gallery'}
-                />
-              </ThemeProvider>
-            </ImageGalleryProvider>
+                  <MessageOverlay<At, Ch, Co, Ev, Me, Re, Us>
+                    overlayOpacity={overlayOpacity}
+                    visible={overlay === 'message'}
+                  />
+                  <ImageGallery<At, Ch, Co, Ev, Me, Re, Us>
+                    imageGalleryCustomComponents={imageGalleryCustomComponents}
+                    overlayOpacity={overlayOpacity}
+                    visible={overlay === 'gallery'}
+                  />
+                  <AttachmentPicker
+                    ref={bottomSheetRef}
+                    {...attachmentPickerProps}
+                  />
+                </ThemeProvider>
+              </ImageGalleryProvider>
+            </AttachmentPickerProvider>
           </MessageOverlayProvider>
         </ChannelInfoOverlayProvider>
       </OverlayContext.Provider>
