@@ -3,23 +3,15 @@ import Svg, { Path, PathProps, SvgProps } from 'react-native-svg';
 
 export type IconProps = Partial<SvgProps> &
   Omit<RootPathProps, 'd'> & {
-    backgroundFill?: SvgProps['fill'];
     height?: number;
     width?: number;
   };
 
 export const RootSvg: React.FC<IconProps> = (props) => {
-  const {
-    backgroundFill = 'none',
-    children,
-    height = 24,
-    viewBox = '0 0 24 24',
-    width = 24,
-  } = props;
+  const { children, height = 24, viewBox = '0 0 24 24', width = 24 } = props;
   return (
     <Svg
       {...{
-        fill: backgroundFill,
         height,
         viewBox,
         width,
