@@ -17,6 +17,17 @@ export let BlurView: BlurView = fail;
 type DeleteFile = ({ uri }: { uri: string }) => Promise<boolean> | never;
 export let deleteFile: DeleteFile = fail;
 
+type GetPhotos = ({
+  after,
+  first,
+}: {
+  first: number;
+  after?: string;
+}) =>
+  | Promise<{ assets: string[]; endCursor: string; hasNextPage: boolean }>
+  | never;
+export let getPhotos: GetPhotos = fail;
+
 type NetInfo = {
   addEventListener: (
     listener: (isConnected: boolean) => void,
@@ -28,15 +39,6 @@ export let NetInfo: NetInfo = {
   addEventListener: fail,
   fetch: fail,
 };
-
-type PickImage = ({
-  compressImageQuality,
-  maxNumberOfFiles,
-}: {
-  compressImageQuality?: number;
-  maxNumberOfFiles?: number;
-}) => Promise<{ cancelled: boolean; images?: { uri: string }[] }> | never;
-export let pickImage: PickImage = fail;
 
 type PickDocument = ({
   maxNumberOfFiles,
@@ -69,6 +71,17 @@ type ShareOptions = {
 type ShareImage = (options: ShareOptions) => Promise<boolean> | never;
 export let shareImage: ShareImage = fail;
 
+type Photo =
+  | {
+      cancelled: false;
+      height: number;
+      uri: string;
+      width: number;
+    }
+  | { cancelled: true };
+type TakePhoto = () => Promise<Photo> | never;
+export let takePhoto: TakePhoto = fail;
+
 type HapticFeedbackMethod =
   | 'selection'
   | 'impactLight'
@@ -83,11 +96,12 @@ export let triggerHaptic: TriggerHaptic = fail;
 type Handlers = {
   BlurView?: BlurView;
   deleteFile?: DeleteFile;
+  getPhotos?: GetPhotos;
   NetInfo?: NetInfo;
   pickDocument?: PickDocument;
-  pickImage?: PickImage;
   saveFile?: SaveFile;
   shareImage?: ShareImage;
+  takePhoto?: TakePhoto;
   triggerHaptic?: TriggerHaptic;
 };
 
@@ -100,12 +114,12 @@ export const registerNativeHandlers = (handlers: Handlers) => {
     deleteFile = handlers.deleteFile;
   }
 
-  if (handlers.NetInfo) {
-    NetInfo = handlers.NetInfo;
+  if (handlers.getPhotos) {
+    getPhotos = handlers.getPhotos;
   }
 
-  if (handlers.pickImage) {
-    pickImage = handlers.pickImage;
+  if (handlers.NetInfo) {
+    NetInfo = handlers.NetInfo;
   }
 
   if (handlers.pickDocument) {
@@ -118,6 +132,10 @@ export const registerNativeHandlers = (handlers: Handlers) => {
 
   if (handlers.shareImage) {
     shareImage = handlers.shareImage;
+  }
+
+  if (handlers.takePhoto) {
+    takePhoto = handlers.takePhoto;
   }
 
   if (handlers.triggerHaptic) {
